% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nfl_athlete_gamelog.R,
%   R/fetch_nfl_athlete_overview.R
\name{create_athlete_gamelog_dataset}
\alias{create_athlete_gamelog_dataset}
\title{Create athlete game log data frame from gamelog response}
\usage{
create_athlete_gamelog_dataset(data, athlete_id)

create_athlete_gamelog_dataset(data, athlete_id)
}
\arguments{
\item{data}{Raw JSON response from ESPN Web API athlete overview endpoint}

\item{athlete_id}{Athlete ID used in request}
}
\value{
data.frame with athlete game log information

Clean data frame with athlete game log information
}
\description{
Processes raw JSON response from ESPN Web API into a structured data frame
containing game-by-game statistics for an NFL athlete
}
\keyword{internal}
