% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nfl_athlete_gamelog.R
\name{create_athlete_season_summary_dataset}
\alias{create_athlete_season_summary_dataset}
\title{Create athlete season summary data frame from gamelog response}
\usage{
create_athlete_season_summary_dataset(data, athlete_id)
}
\arguments{
\item{data}{Raw JSON response from ESPN Web API athlete gamelog endpoint}

\item{athlete_id}{Character. Athlete ID used in request}
}
\value{
data.frame with athlete season summary information
}
\description{
Processes raw JSON response to create season summary statistics
}
\keyword{internal}
