% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nfl_athlete_splits.R
\name{create_athlete_splits_dataset}
\alias{create_athlete_splits_dataset}
\title{Create athlete splits data frame from API response}
\usage{
create_athlete_splits_dataset(data, athlete_id)
}
\arguments{
\item{data}{Raw JSON response from ESPN Web API athlete splits endpoint}

\item{athlete_id}{Character. Athlete ID used in request}
}
\value{
data.frame with athlete splits information
}
\description{
Processes raw JSON response from ESPN Web API into a structured data frame
containing splits data for an NFL athlete
}
\keyword{internal}
