% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_college_football_team_news.R
\name{create_cfb_news_dataset}
\alias{create_cfb_news_dataset}
\title{Create college football news data frame from Site API response}
\usage{
create_cfb_news_dataset(data)
}
\arguments{
\item{data}{Raw JSON response from ESPN Site API news endpoint}
}
\value{
Data frame with news articles information
}
\description{
Processes raw JSON response from ESPN Site API into structured data frame
containing college football news information
}
\keyword{internal}
