% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_college_football_scoreboard.R
\name{create_cfb_scoreboard_dataset}
\alias{create_cfb_scoreboard_dataset}
\title{Create college football scoreboard data frame from Site API response}
\usage{
create_cfb_scoreboard_dataset(data, groups = NA, week = NA, dates = NA)
}
\arguments{
\item{data}{Raw JSON response from ESPN Site API scoreboard endpoint}

\item{groups}{Character. Group IDs used in request (if any)}

\item{week}{Character. Week number used in request (if any)}

\item{dates}{Character. Date used in request (if any)}
}
\value{
Data frame with scoreboard information
}
\description{
Processes raw JSON response from ESPN Site API into structured data frame
containing college football scoreboard information
}
\keyword{internal}
