% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_college_football_playbyplay.R
\name{create_cfb_scoring_plays_dataset}
\alias{create_cfb_scoring_plays_dataset}
\title{Create college football scoring plays data frame from Site API response}
\usage{
create_cfb_scoring_plays_dataset(data, game_id)
}
\arguments{
\item{data}{Raw JSON response from ESPN Site API game summary endpoint}

\item{game_id}{Character. Game ID used in request}
}
\value{
Data frame with scoring plays information
}
\description{
Processes raw JSON response from ESPN Site API into structured data frame
containing detailed scoring plays information
}
\keyword{internal}
