% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nfl_athletes.R
\name{create_clean_athletes_dataset}
\alias{create_clean_athletes_dataset}
\title{Create clean athletes dataset from ESPN Core API response}
\usage{
create_clean_athletes_dataset(data, params)
}
\arguments{
\item{data}{Raw JSON response from ESPN Core API athletes endpoint}

\item{params}{Query parameters used in request}
}
\value{
Clean data frame with athlete information
}
\description{
Create clean athletes dataset from ESPN Core API response
}
