% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nfl_athletes.R
\name{create_filter_message}
\alias{create_filter_message}
\title{Create informative filter message for console output}
\usage{
create_filter_message(params)
}
\arguments{
\item{params}{Query parameters used in request}
}
\value{
Character string describing applied filters
}
\description{
Create informative filter message for console output
}
