% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mens_college_basketball_boxscore.R,
%   R/fetch_nba_boxscore.R, R/fetch_wnba_boxscore.R,
%   R/fetch_womens_college_basketball_boxscore.R
\name{create_player_stats_dataset}
\alias{create_player_stats_dataset}
\title{Create player statistics data frame from ESPN boxscore API}
\usage{
create_player_stats_dataset(data, event_id)

create_player_stats_dataset(data, event_id)

create_player_stats_dataset(data, event_id)

create_player_stats_dataset(data, event_id)
}
\arguments{
\item{data}{Raw JSON response from ESPN boxscore API}

\item{event_id}{Event ID used in the request}
}
\value{
Data frame with player statistics

Data frame with player statistics

Data frame with player statistics

Data frame with player statistics
}
\description{
Processes player statistics from ESPN boxscore players data

Processes player statistics from ESPN boxscore players data

Processes player statistics from ESPN boxscore players data

Processes player statistics from ESPN boxscore players data
}
\keyword{internal}
