% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_soccer_league_squads.R
\name{create_soccer_squad_filter_message}
\alias{create_soccer_squad_filter_message}
\title{Create informative filter message for squad}
\usage{
create_soccer_squad_filter_message(league, team_id, data)
}
\arguments{
\item{league}{League identifier}

\item{team_id}{Team ID}

\item{data}{Raw API response (to extract team name)}
}
\value{
Character string describing applied filters
}
\description{
Create informative filter message for squad
}
\keyword{internal}
