% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_soccer_transfers.R
\name{create_transfers_filter_message}
\alias{create_transfers_filter_message}
\title{Create informative filter message for transfers}
\usage{
create_transfers_filter_message(league)
}
\arguments{
\item{league}{League identifier}
}
\value{
Character string describing applied filters
}
\description{
Create informative filter message for transfers
}
\keyword{internal}
