% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_wnba_teams.R
\name{create_wnba_teams_dataset}
\alias{create_wnba_teams_dataset}
\title{Create wnba teams data frame from Site API response}
\usage{
create_wnba_teams_dataset(data)
}
\arguments{
\item{data}{Raw JSON response from ESPN Site API teams endpoint}
}
\value{
Data frame with teams information
}
\description{
Processes raw JSON response from ESPN Site API into structured data frame
containing wnba teams information
}
\keyword{internal}
