% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_womens_college_basketball_team_news.R
\name{create_womens_college_basketball_news_dataset}
\alias{create_womens_college_basketball_news_dataset}
\title{Create womens_college_basketball team news data frame from Site API response}
\usage{
create_womens_college_basketball_news_dataset(data, team_id)
}
\arguments{
\item{data}{Raw JSON response from ESPN Site API news endpoint}

\item{team_id}{Character. Team ID used in request}
}
\value{
Data frame with news articles information
}
\description{
Processes raw JSON response from ESPN Site API into structured data frame
containing team-specific news information
}
\keyword{internal}
