% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_nfl_standings.R
\name{fetch_nfl_standings}
\alias{fetch_nfl_standings}
\title{Fetch actual NFL standings data}
\usage{
fetch_nfl_standings(
  season = as.numeric(format(Sys.Date(), "\%Y")),
  raw = FALSE
)
}
\arguments{
\item{season}{Numeric. Season year (e.g., 2023, 2024). Default current year.}

\item{raw}{Logical. If TRUE, assigns raw JSON to global environment. Default FALSE.}
}
\value{
Invisibly returns the data, but main purpose is global environment assignment.
}
\description{
Fetch actual NFL standings data
}
\examples{
# Get current regular season standings
fetch_nfl_standings(season = 2024)
head(nfl_standings)

# Get raw data for custom processing
fetch_nfl_standings(season = 2024, raw = TRUE)
str(nfl_standings_raw, max.level = 3)

}
