% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mens_college_basketball_player_stats.R,
%   R/fetch_nba_player_stats.R, R/fetch_wnba_player_stats.R,
%   R/fetch_womens_college_basketball_player_stats.R
\name{fetch_player_info}
\alias{fetch_player_info}
\title{Get player information from ESPN player profile API}
\usage{
fetch_player_info(player_id)

fetch_player_info(player_id)

fetch_player_info(player_id)

fetch_player_info(player_id)
}
\arguments{
\item{player_id}{ESPN player ID}
}
\value{
List with player information

List with player information

List with player information

List with player information
}
\description{
Since the gamelog API doesn't include athlete info, fetch it separately

Since the gamelog API doesn't include athlete info, fetch it separately

Since the gamelog API doesn't include athlete info, fetch it separately

Since the gamelog API doesn't include athlete info, fetch it separately
}
\keyword{internal}
