% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_college_football_rankings.R
\name{get_cfb_ranking_polls}
\alias{get_cfb_ranking_polls}
\title{Get college football ranking poll types}
\usage{
get_cfb_ranking_polls()
}
\value{
Named list with poll information
}
\description{
Returns information about available ranking polls
}
